+++ samples/sampler/stories/default/ProgressBar.js
@@ -22,9 +22,12 @@ storiesOf('Sandstone', module)
 			const tooltip = boolean('tooltip', ProgressBarTooltipConfig);
 			const position = select('position', ['', 'above', 'above left', 'above center', 'above right', 'above before', 'above after', 'before', 'left', 'right', 'after', 'below', 'below left', 'below center', 'below right', 'below before', 'below after'], ProgressBarTooltipConfig, '');
 
+			const balanced = boolean('balanced', ProgressBarConfig);
+
 			return (
 				<ProgressBar
 					backgroundProgress={number('backgroundProgress', ProgressBarConfig, {range: true, min: 0, max: 1, step: 0.01}, 0.5)}
+					balanced={balanced}
 					disabled={disabled}
 					highlighted={boolean('highlighted', ProgressBarConfig)}
 					orientation={select('orientation', ['horizontal', 'vertical', 'radial'], ProgressBarConfig, 'horizontal')}